# swagger/swagger.yml
openapi: 3.0.0
info:
  title: 'UMC Study API'
  version: '1.0.0'
  description: 'UMC Study API with express, API 설명'
servers:
  - url: http://localhost:3000

paths:
  /mission/add-mission:
    post:
      tags:
        - Mission
      summary: Add a new mission
      description: Add a new mission to the store
      parameters:
        - in: query
          name: storeId
          schema:
            type: integer
          description: The ID of the store
          example: 1
        - in: query
          name: reward
          schema:
            type: integer
          description: The reward for the mission
          example: 100
        - in: query
          name: deadline
          schema:
            type: string
            format: date-time
          description: The deadline of the mission
          example: 2024-12-31T23:59:59.999Z
        - in: query
          name: missionSpec
          schema:
            type: string
          description: The specification of the mission
          example: Sample mission spec
      responses:
        '200':
          description: 미션 추가 성공!
          schema:
            type: object
            properties:
              status:
                type: integer
                example: 200
              isSuccess:
                type: boolean
                example: true
              code:
                type: integer
                example: 2000
              message:
                type: string
                example: "success!"
              data:
                type: object
                example: 
                    "storeId": storeId,
                    "reward": reward,
                    "deadline": deadline,
                    "missionSpec": missionSpec
        '400':
          description: 잘못된 요청
          schema:
            type: object
            properties:
              status:
                type: integer
                example: 400
              isSuccess:
                type: boolean
                example: false
              code:
                type: integer
                example: COMMON001
              message:
                type: string
                example: 잘못된 요청입니다

        '500':
          description: 서버 에러
          schema:
            type: object
            properties:
              status:
                type: integer
                example: 500
              isSuccess:
                type: boolean
                example: false
              code:
                type: integer
                example: COMMON000
              message:
                type: string
                example: 서버 에러, 관리자에게 문의 바랍니다.
  /mission/member-mission:
    post:
      tags:
        - MemberMission
      summary: Add a new member mission
      description: Add a new mission for a member
      parameters:
        - in: query
          name: memberId
          schema:
            type: integer
          description: The ID of the member
          example: 1
        - in: query
          name: missionId
          schema:
            type: integer
          description: The ID of the mission
          example: 1
      responses:
        '200':
          description: 미션 추가 성공!
          schema:
            type: object
            properties:
              status:
                type: integer
                example: 200
              isSuccess:
                type: boolean
                example: true
              code:
                type: integer
                example: 2000
              message:
                type: string
                example: "success!"
              data:
                type: object
                example: 
                    "memberId": memberId,
                    "missionId": missionId,
                    "status": status
                
                
        '400':
          description: 잘못된 요청
          schema:
            type: object
            properties:
              status:
                type: integer
                example: 400
              isSuccess:
                type: boolean
                example: false
              code:
                type: integer
                example: COMMON001
              message:
                type: string
                example: 잘못된 요청입니다

        '500':
          description: 서버 에러
          schema:
            type: object
            properties:
              status:
                type: integer
                example: 500
              isSuccess:
                type: boolean
                example: false
              code:
                type: integer
                example: COMMON000
              message:
                type: string
                example: 서버 에러, 관리자에게 문의 바랍니다.
      
    